:recycle: search repository refactor
